Miktar bilgisi decimal türüne çevirildi veritabanı ve view dosyalarında(admin-ithalat evrak oluşturma için )

+++ database/migrations/2025_03_05_134256_create_saglik_sertifikas_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
 
             $table->string('ssn');
-            $table->integer('toplam_miktar');
-            $table->integer('kalan_miktar');
+            $table->decimal('toplam_miktar',10,3);
+            $table->decimal('kalan_miktar',10,3);
 
             $table->timestamps();
         });
